# docs/swagger.yaml
basePath: /
definitions:
...
  model.AnswerOptionDto:
    properties:
      content:
        type: string
      id:
        type: string
      isCorrect:
        type: boolean
    required:
    - content
    type: object
...
  model.CreateAnswerOptionDto:
    properties:
      content:
        type: string
      isCorrect:
        type: boolean
    required:
    - content
    - isCorrect
    type: object
  model.CreateQuestionBatchItem:
    properties:
      content:
        type: string
      options:
        items:
          $ref: '#/definitions/model.CreateAnswerOptionDto'
        minItems: 1
        type: array
      timeLimit:
        minimum: 1
        type: integer
    required:
    - content
    - options
    - timeLimit
    type: object
  model.CreateQuestionBatchRequest:
    properties:
      level:
        enum:
        - easy
        - medium
        - hard
        type: string
      questions:
        items:
          $ref: '#/definitions/model.CreateQuestionBatchItem'
        minItems: 1
        type: array
      topicId:
        type: string
    required:
    - level
    - questions
    - topicId
    type: object
  model.GeneralResponse:
    properties:
      code:
        type: integer
      data: {}
      errors: {}
      message:
        type: string
    type: object
...
  model.QuestionResponse:
    properties:
      content:
        type: string
      createdAt:
        type: string
      id:
        type: string
      level:
        type: string
      options:
        items:
          $ref: '#/definitions/model.AnswerOptionDto'
        type: array
      timeLimit:
        type: integer
      topicId:
        type: string
      updatedAt:
        type: string
    type: object
...
  model.UpdateQuestionRequest:
    properties:
      content:
        type: string
      options:
        items:
          $ref: '#/definitions/model.AnswerOptionDto'
        minItems: 1
        type: array
      timeLimit:
        minimum: 1
        type: integer
    required:
    - content
    - options
    - timeLimit
    type: object
...
host: localhost:9999
info:
  contact:
    email: fiber@swagger.io
    name: API Support
  description: Baseline project using Go Fiber
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Go Fiber Clean Architecture
  version: 1.0.0
paths:
...
  /api/v1/questions:
    get:
      consumes:
      - application/json
      description: list all questions.
      parameters:
      - description: Filter by topic ID
        in: query
        name: topicId
        required: true
        type: string
      - description: Page number
        in: query
        name: page
        type: integer
      - description: Limit number
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/model.GeneralResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/model.QuestionResponse'
                  type: array
              type: object
      summary: list questions
      tags:
      - Question
  /api/v1/questions/{id}:
    delete:
      consumes:
      - application/json
      description: delete a question by UUID.
      parameters:
      - description: Question ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/model.GeneralResponse'
            - properties:
                data:
                  type: string
              type: object
      security:
      - JWT: []
      summary: delete question
      tags:
      - Question
    get:
      consumes:
      - application/json
      description: get a specific question by UUID.
      parameters:
      - description: Question ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/model.GeneralResponse'
            - properties:
                data:
                  $ref: '#/definitions/model.QuestionResponse'
              type: object
      summary: get question
      tags:
      - Question
    put:
      consumes:
      - application/json
      description: update an existing question and sync its answers.
      parameters:
      - description: Question ID
        in: path
        name: id
        required: true
        type: string
      - description: Update Question Request Body
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/model.UpdateQuestionRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/model.GeneralResponse'
            - properties:
                data:
                  $ref: '#/definitions/model.QuestionResponse'
              type: object
      security:
      - JWT: []
      summary: update question
      tags:
      - Question
  /api/v1/questions/batch:
    post:
      consumes:
      - application/json
      description: create questions in batch for a topic.
      parameters:
      - description: Create Questions Batch Request Body
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/model.CreateQuestionBatchRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/model.GeneralResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/model.QuestionResponse'
                  type: array
              type: object
      security:
      - JWT: []
      summary: create questions batch
      tags:
      - Question
...
schemes:
- http
- https
securityDefinitions:
  JWT:
    description: Authorization For JWT
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"